初始提交

+++ database/migrations/2018_10_06_194449_create_complexes_table.php
@@ -17,7 +17,7 @@ class CreateComplexesTable extends Migration
             $table->increments('id');
             $table->integer('user_id')->unsigned()->comment('外键：users的主键');
             $table->integer('history_read_count')->default(0)->comment('历史浏览次数');
-            $table->decimal('history_amount', 10,2)->default(0)->comment('历史分润金额');
+            $table->integer('history_amount')->default(0)->comment('历史分润金额');
             $table->timestamps();
 
             $table->foreign('user_id')
